# src/docs/yaml/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Back End Demonstration API
  description: A simple blog API.
  contact:
    name: Ko Wing To
    email: 227108934@stu.vtc.edu.hk
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'https://backend.koto123.repl.co/api'
    description: Development server
...
paths:
...
  /addchatrooms:
    post:
      summary: Create a new chatroom
      tags:
        - Chatrooms
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: './chatroom.json'
        required: true
      responses:
        '201':
          description: Chatroom created
          content:
            application/json:
              schema:
                $ref: './chatroom.json'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /chatrooms:
    get:
      summary: Get all chatrooms
      tags:
        - Chatrooms
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: './chatroom.json'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /chatrooms/{id}:
    get:
      summary: Get a chatroom by ID
      tags:
        - Chatrooms
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the chatroom to retrieve
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: './chatroom.json'
        '404':
          description: Chatroom not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /chatrooms/user/{userId}:
    get:
      summary: Get a chatroom by userID
      tags:
        - Chatrooms
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
          description: ID of the user to retrieve chatrooms for
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: './chatroom.json'
        '404':
          description: Chatroom not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /messages:
    post:
      summary: Create a new message
      tags:
        - Messages
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: './message.json'
        required: true
      responses:
        '201':
          description: Message created
          content:
            application/json:
              schema:
                $ref: './message.json'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /messages/{chatroomId}:
    get:
      summary: Get all messages by chatroom
      tags:
        - Messages
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: chatroomId
          schema:
            type: string
          required: true
          description: ID of the chatroom to retrieve messages for
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: './message.json'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /messages/{id}:
    delete:
      summary: Delete a message by ID
      tags:
        - Messages
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the message to delete
      responses:
        '200':
          description: Message deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
components:
...
  securitySchemes:
      bearerAuth:          
        type: http
        scheme: bearer
        bearerFormat: JWT   